requirements.txt
allure-pytest==2.13.2
pytest==7.3.1
selene==2.0.0rc4
selenium==4.10.0
webdriver_manager==3.8.6
testit-importer-allure==1.3.3
testit-api-client==3.0.1
python-dotenv==0.20.0
requests==2.31.0